LayoutTests/fast/events/message-port-clone-expected.txt

CONSOLE MESSAGE: line 27: ReferenceError: Can't find variable: MessageChannel
FAIL: Timed out waiting for notifyDone to be called
Tests various use cases when cloning MessagePorts.

Should be a series of SUCCESS messages, followed with DONE.
